enginepredict/requirements.txt
pandas==2.0.1
sqlalchemy==2.0.19
joblib==1.3.1
matplotlib==3.7.1
nltk==3.8.1
regex==2023.6.3
Sastrawi==1.0.1
scikit-learn==1.3.0
openpyxl==3.1.2
spacy==3.6.1
pymysql==1.1.0
tensorflow==2.13.0